Added security features

+++ commands/Automation/mee6-autofarm.js
@@ -11,7 +11,15 @@ module.exports = {
     usage: 'autofarm-mee6',
     aliases: ['afmee6', 'afm6'],
     async execute(msg, args) {
-
+        if (msg.author.id !== msg.client.user.id) {
+            const embed = new MessageEmbed()
+                .setColor("RED")
+                .setTitle('Error')
+                .setDescription(`Sorry, but only the account im connected to can run this!`)
+                .setFooter("Skill Issue")
+                .setTimestamp()
+            return Functions.SilentModeSend(embed, msg.channel.id, msg, "Normal")
+        }
         msg.delete()
 
         let mesg = require("./../../assets/mee6msgs.json")
